requirements.txt
torch
transformers
sentence-transformers
chromadb
keybert
easyocr
opencv-python
pillow
langchain
langchain-community
langchain-core
tqdm